Cleaned Up Code
